+++ sdk/src/admin/index.ts
@@ -1,5 +1,6 @@
 export * from "./deployScripts.js";
 export * from "./registerTreasuryStake.js";
 export * from "./verifyDeployment.js";
+export * from "./verifyProtocolDeployment.js";
 export * from "./initializeSettings.js";
 export * from "./verifySettings.js";